add initial startup library, move all code from player/plays3m.pas to player/plays3m.c, small C string manipulation fixes

+++ src/common.c
@@ -12,7 +12,7 @@
 
 #include "common.h"
 
-void *__new(uint16_t size)
+void *__new(size_t size)
 {
     uint16_t seg;
